Merging Integration Tests

+++ src/main/java/au/edu/rmit/sept/SuperPrice/controller/CustomProductPriceController.java
@@ -13,9 +13,6 @@ import java.util.List;
  * CustomController class handles requests for custom queries
  * Base URL: http://localhost:8080/SuperPrice/v2/ProductPrices
  * CrossOrigin: http://localhost:5173
- * 
- * TODO: Update final URLs
- * TODO: Move test controller methods to tests & adapt as needed
  */
 @RestController
 @RequestMapping(value = "/SuperPrice/v2/ProductPrices")
@@ -24,38 +21,6 @@ public class CustomProductPriceController {
     // Declare custom service instance
     @Autowired
     private CustomProductPriceService customService;
-
-    // TODO: Test this method / remove if not used
-    // Get custom ProductPrices data by selected inputs or display all for default supermarket if none selected (related Products, ProductPrices, ProductRewards & Supermarkets data)
-    // @GetMapping
-    // public ResponseEntity<List<CustomDTOProductPrices>> getProductsBySupermarketIdAndProductCategory(@RequestParam(required = false) Long supermarket_id, @RequestParam(required = false) String product_category) {
-    //     if (supermarket_id == null && product_category == null) {
-    //         // Get all ProductPrices for default supermarket (supermarket_id = 1)
-    //         int supermarketId = 1;
-    //         List<CustomDTOProductPrices> customProductPriceData = this.customService.getProductPricesBySupermarketId(supermarketId);
-
-    //         // Convert to ResponseEntity
-    //         if (!customProductPriceData.isEmpty()) {
-    //             // OK
-    //             return ResponseEntity.ok(customProductPriceData);
-    //         } else {
-    //             // No Content
-    //             return ResponseEntity.noContent().build();
-    //         }
-    //     } else {
-    //         // Get all ProductPrices in selected category & supermarket
-    //         List<CustomDTOProductPrices> customProductPriceData = this.customService.getProductPricesBySupermarketIdAndProductCategory(supermarket_id, product_category);
-
-    //         // Convert to ResponseEntity
-    //         if (!customProductPriceData.isEmpty()) {
-    //             // OK
-    //             return ResponseEntity.ok(customProductPriceData);
-    //         } else {
-    //             // No Content
-    //             return ResponseEntity.noContent().build();
-    //         }
-    //     }
-    // }
 
     // Get custom ProductPrices data by given product_id (related Products, ProductPrices, ProductRewards & Supermarkets data)
     @GetMapping("/{product_id}")
@@ -107,7 +72,7 @@ public class CustomProductPriceController {
     // Get custom ProductPrices data by selected supermarket_id & product_category (related Products, ProductPrices, ProductRewards & Supermarkets data)
     // Test with http://localhost:8080/v2/ProductPrices/Supermarket/1/Dairy Products
     @GetMapping("/Supermarket/Category/{supermarket_id}/{product_category}")
-    public ResponseEntity<List<CustomDTOProductPrices>> getProductPricesBySupermarketIdAndProductCategory(@PathVariable("supermarket_id") Long supermarket_id, @PathVariable("product_category") String product_category) {
+    public ResponseEntity<List<CustomDTOProductPrices>> getProductPricesBySupermarketIdAndProductCategory(@PathVariable("supermarket_id") int supermarket_id, @PathVariable("product_category") String product_category) {
         List<CustomDTOProductPrices> customProductPriceData = this.customService.getProductPricesBySupermarketIdAndProductCategory(supermarket_id, product_category);
 
         // Convert to ResponseEntity
@@ -119,9 +84,4 @@ public class CustomProductPriceController {
             return ResponseEntity.noContent().build();
         }
     }
-
-
-    // TODO: Add any additional custom controller methods here:
-
-
 }
